# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "backend api with go",
        "contact": {}
    },
    "paths": {
        "/tasks": {
            "get": {
                "description": "Get all tasks from the system.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all tasks.",
                "operationId": "get-all-tasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/tasks_domain.Task"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new task with the data provided in the JSON body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new task",
                "parameters": [
                    {
                        "description": "Data of the new task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasks_domain.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inserted task",
                        "schema": {
                            "$ref": "#/definitions/tasks_domain.Task"
                        }
                    },
                    "400": {
                        "description": "Error in input data",
                        "schema": {
                            "$ref": "#/definitions/tasks_domain.Task"
                        }
                    },
                    "500": {
                        "description": "Error in input data",
                        "schema": {
                            "$ref": "#/definitions/tasks_domain.Task"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "put": {
                "description": "Update a task in the system using its ID and the new provided data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a task by ID.",
                "operationId": "update-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the task to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New data of the task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasks_domain.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a task from the system using its ID.",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a task by ID.",
                "operationId": "delete-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the task to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "tasks_domain.Task": {
            "type": "object",
            "required": [
                "name",
                "status"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}